# openapi/operator.yaml
openapi: 3.0.3
info:
  title: Karta operator API
  version: 1.0.0
  description: |
    Publication status, snapshot digest authorization, activation, rollback
    and cleanup. Served by the publisher (`karta publisher`) on its own
    listener (`KARTA_OPERATOR_LISTEN_ADDR`, published on 127.0.0.1:8081 by
    default); the public API process has no operator routes and no write
    access to the registry. See docs/security.md for the threat model and
    docs/runbook.md for procedures.

    **Authentication.** Every `/v1/operator` request needs
    `Authorization: Bearer <token>`, a 256-bit random token (64 lowercase hex
    digits) whose SHA-256 is listed in the publisher's credentials file with
    a name and scopes: `status` (read status and audit), `publish` (authorize
    or revoke digests, activate a ready release), `rollback`, `cleanup`.
    Missing or wrong tokens get `401 unauthenticated`, a missing scope
    `403 forbidden`; both are logged and audited (at most 30 audit rows per
    minute for refusals). Tokens are never logged.

    **Requests.** `POST` bodies are `application/json`, at most 16 KiB, one
    object without unknown fields. Every action requires a `reason`
    (1–500 characters, no control characters), which is recorded in the
    audit log with the credential name, action, target and outcome. Every
    request is bounded by `KARTA_OPERATOR_REQUEST_TIMEOUT`. Responses are
    `Cache-Control: no-store` and carry `X-Request-ID`.

...
  license:
...
servers:
...
security:
  - bearer: []
paths:
...
  /v1/operator/status:
    get:
      summary: Publication status
      operationId: operatorStatus
      description: Scope `status`. Region and policy, active and recent releases, recent submissions, open digest authorizations, the inbox, the job in progress and storage use.
      responses:
        "200":
          description: Status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/operator/audit:
    get:
      summary: Audit log, newest first
      operationId: operatorAudit
      description: Scope `status`. Append-only records of every publication step and operator action.
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
        - { name: before_id, in: query, description: Only records with a smaller id (paging).,
            schema: { type: integer, format: int64, minimum: 1 } }
      responses:
        "200":
          description: Audit records.
          content:
            application/json:
              schema:
                type: object
                required: [entries]
                properties:
                  entries: { type: array, items: { $ref: "#/components/schemas/AuditEntry" } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/operator/openapi.yaml:
    get:
      summary: This document
      operationId: operatorOpenAPI
      description: Scope `status`.
      responses:
        "200":
          description: OpenAPI document.
          content:
            application/yaml:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /v1/operator/authorizations:
    post:
      summary: Authorize one exact snapshot digest for the configured region
      operationId: operatorAuthorize
      description: |
        Scope `publish`. Records that the operator has verified this exact
        snapshot (by SHA-256, and size if given) out of band. Only a digest
        that is pinned in the region configuration or authorized here is
        imported; there is no way to accept arbitrary files. An open,
        unexpired authorization for the same digest and size is returned
        unchanged (`200`, `created: false`). A submission rejected earlier as
        `unauthorized_digest` is re-evaluated on the next inbox scan once its
        digest is authorized.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [sha256, reason]
              properties:
                sha256: { $ref: "#/components/schemas/SHA256" }
                size_bytes: { type: integer, format: int64, minimum: 1, description: "If given, only a file of exactly this size matches." }
                expires_at: { type: string, format: date-time, description: "Optional end of the authorization, in the future and at most one year ahead." }
                reason: { $ref: "#/components/schemas/Reason" }
      responses:
        "201":
          description: Authorization created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthorizationResult" }
        "200":
          description: An equal authorization already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthorizationResult" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/operator/authorizations/{sha256}/revoke:
    post:
      summary: Revoke the open authorization of a digest
      operationId: operatorRevoke
      description: Scope `publish`. Releases already built from the digest are not affected.
      parameters:
        - { name: sha256, in: path, required: true, schema: { $ref: "#/components/schemas/SHA256" } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReasonBody" }
      responses:
        "200":
          description: Whether an open authorization was revoked.
          content:
            application/json:
              schema:
                type: object
                required: [revoked]
                properties:
                  revoked: { type: boolean }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/operator/releases/{release_id}/activate:
    post:
      summary: Activate a ready or retired release (forward only)
      operationId: operatorActivate
      description: |
        Scope `publish`. The target must be retained and validated (`ready`
        or `retired`), pass the same compatibility checks as serving (schema
        major, style, serving toolchain), and satisfy the forward rule: a
        newer data timestamp than the active release, or the same snapshot
        rebuilt. Going back is a rollback. The previous release stays servable
        to pinned clients for the pin grace period.
      parameters:
        - { name: release_id, in: path, required: true, schema: { $ref: "#/components/schemas/ReleaseID" } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SwitchBody" }
      responses:
        "200":
          description: Switched, or already active (`changed` false).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SwitchResult" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/operator/rollback:
    post:
      summary: Roll back to a retained, validated, compatible release
      operationId: operatorRollback
      description: |
        Scope `rollback`. Without `release_id`, the target is the most
        recently replaced retained (`retired`) release. The target must be
        `ready` or `retired` (never `removing`, `removed` or `failed`) and
        pass the serving compatibility checks. Only the active pointer
        changes; release contents are immutable. A region change must be
        explicit (`allow_region_change`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SwitchBody" }
      responses:
        "200":
          description: Rolled back, or the target is already active (`changed` false).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SwitchResult" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/operator/cleanup:
    post:
      summary: Remove releases beyond the retention policy
      operationId: operatorCleanup
      description: |
        Scope `cleanup`. Keeps the active release, the
        `KARTA_RETAIN_RELEASES` most recent other validated releases, and
        every release still inside its pin grace (plus the drain margin).
        A release whose database has any session is skipped. `dry_run` only
        reports the decisions. Cleanup also runs after each publication and
        every `KARTA_CLEANUP_INTERVAL`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: { $ref: "#/components/schemas/Reason" }
                dry_run: { type: boolean, default: false }
      responses:
        "200":
          description: Decisions and what was removed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CleanupResult" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  responses:
    Error:
      description: Error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    ReleaseID: { type: string, pattern: "^r[0-9a-f]{24}$" }
    SHA256: { type: string, pattern: "^[0-9a-f]{64}$" }
    Reason: { type: string, minLength: 1, maxLength: 500, description: Why; recorded in the audit log. }
    ReasonBody:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { $ref: "#/components/schemas/Reason" }
    Timestamp: { type: string, format: date-time }
...
    AuthorizationResult:
      type: object
      required: [authorization, created]
      properties:
        authorization: { $ref: "#/components/schemas/Authorization" }
        created: { type: boolean }
    AuditEntry:
      type: object
      required: [id, at, actor, source, action, outcome]
      properties:
        id: { type: integer, format: int64 }
        at: { $ref: "#/components/schemas/Timestamp" }
        actor: { type: string, description: "Credential name, `inbox`, `cli`, `system` or `anonymous`." }
        source: { type: string, enum: [operator_api, inbox, cli, system] }
        action: { type: string }
        target: { type: string }
        outcome: { type: string, enum: [succeeded, rejected, failed, denied, noop] }
        reason: { type: string }
        request_id: { type: string }
        detail: {}
...
    Status:
      type: object
      required: [region, policy, active, releases, submissions, authorizations, inbox, job, storage]
      properties:
        region:
          type: object
          required: [id, name, pinned_digests, require_provenance]
          properties:
            id: { type: string }
            name: { type: string }
            pinned_digests: { type: array, items: { $ref: "#/components/schemas/SHA256" } }
            require_provenance: { type: boolean }
            error: { type: string }
        policy:
          type: object
          required: [auto_activate, pin_grace_seconds, retain_releases, max_attempts, inbox_settle_seconds]
          properties:
            auto_activate: { type: boolean }
            pin_grace_seconds: { type: number }
            retain_releases: { type: integer }
            max_attempts: { type: integer }
            inbox_settle_seconds: { type: number }
        active: { allOf: [{ $ref: "#/components/schemas/Release" }], nullable: true }
        releases: { type: array, items: { $ref: "#/components/schemas/ReleaseStatus" } }
        submissions: { type: array, items: { $ref: "#/components/schemas/Submission" } }
        authorizations: { type: array, items: { $ref: "#/components/schemas/Authorization" } }
        inbox:
          type: object
          required: [enabled, last_scan]
          properties:
            enabled: { type: boolean }
            last_scan:
              type: object
              required: [at, pending]
              properties:
                at: { allOf: [{ $ref: "#/components/schemas/Timestamp" }], nullable: true }
                error: { type: string }
                pending:
                  type: array
                  items:
                    type: object
                    required: [name, waiting]
                    properties:
                      name: { type: string }
                      waiting: { type: string, enum: [waiting_for_ready_marker, waiting_for_snapshot, settling] }
                      problem: { type: string }
        job:
          type: object
          nullable: true
          required: [submission_id, name, source, phase, started_at, phase_started_at]
          properties:
            submission_id: { type: integer, format: int64 }
            name: { type: string }
            source: { type: string }
            phase: { type: string }
            started_at: { $ref: "#/components/schemas/Timestamp" }
            phase_started_at: { $ref: "#/components/schemas/Timestamp" }
            release_id: { type: string }
        storage:
          type: object
          required: [release_bytes, candidate_estimate_bytes, budget_bytes, db_volume_free_bytes, staging_free_bytes, min_free_bytes]
          properties:
            release_bytes: { type: integer, format: int64 }
            candidate_estimate_bytes: { type: integer, format: int64 }
            budget_bytes: { type: integer, format: int64 }
            db_volume_free_bytes: { type: integer, format: int64, nullable: true }
            staging_free_bytes: { type: integer, format: int64, nullable: true }
            min_free_bytes: { type: integer, format: int64 }
        failpoints: { type: array, items: { type: string }, description: Only present in test runs with fault injection. }